The big thing is the wall dissolver.
Added a wall dissolver item, but have not yet updated the AI to account
for it. Currently the item appears on every maze for debugging reasons,
but it is intended to be super rare because it can make the maze much
easier for human players. Added coding style guidelines. Added some pro
tips. Replaced a bunch of incorrent "auto" type specifiers with
"decltype", which does as I had intended with "auto". Changed a couple
functions in FontManager to take u32s rather than unsigned ints.