Multiple gamepad support added
Hypothetically, the program should now be able to work with any number
of gamepads up to 255, not just a single gamepad like it did before.
Unfortunately I only have one gamepad with which to test it. The game is
still only playable with gamepads in a technical sense, as gamepad
events happen far more often than keyboard or mouse events and so the
player ends up slip-sliding all over the place.

Signed-off-by: James Dearing <dearingj@lifetime.oregonstate.edu>